Add support for pipeline provider defaults
The choice here is to have mandatory defaults that cover for everything a
tekton provider needs and let the pipeline provider redefine whatever it
needs.  In this way we don't need to do hardcode any field in
qontract-reconcile outside of the graphql query as we know that all providers
will always have everything from the defaults.

Part of APPSRE-4165

Signed-off-by: Rafa Porres Molina <rporresm@redhat.com>