Enhance metadata associated with dependencies in app-1.yml
The current implementation of dependencies in app-1.yml doesn't provide
much context about how a dependency is used. For example, is it a build
or runtime dependency? If it is a runtime dependency, knowing whether it
is upstream or downstream provides more information for troubleshooting
and also allows for creating directional graphs of the flow of traffic
across multiple services. A description field allows teams to provide
even more application-specific context which can also be helpful for
SREs that may not be very familiar with a given service.

dependencyFailureImpact was previously associated with a
dependency-1.yml, but multiple apps could be relying on the same
dependency, all with different impacts if the dependency is down. At some
point we may consider deprecating that field from the dependency itself?

A related change to dependency-1.yml also allows linking an app-1.yml
file to the dependency. This should make it easier to track down
information about a dependency when it is an internal app.