add signalfx as endpoint monitoring provider (#94)
following the provider pattern for endpoint monitoring established in the design doc
https://gitlab.cee.redhat.com/service/app-interface/-/blob/master/docs/app-sre/design-docs/service-endpoint-monitoring.md,
this schema change adds signalfx metric scraping from https://github.com/geoberle/signalfx-prometheus-exporter

since signalfx can be a metric store for arbitrary metric producing systems like catchpoint, this
integration will open a way to get ahold of those metric sources in a generic way.

similar to the blackbox exporter provider, this implementation aquires metrics by scraping
an endpoint that returns target specific metrics. while the target for blackbox-exporter is usually
a URL, for signalfx it is a label/value combination for metrics exposed by signalfx cloud.
this is due to the fact that signalfx is not just about URL but more generic about metrics with labels,
similar to promtheus itself.

first lets define the provider. the actual scraping happens via a `Probe` CRD and the
specified namespace will host this CR. additionally, the exporterUrl references the
scrape URL

```yaml
  ---
  $schema: /dependencies/endpoint-monitoring-provider-1.yml
  ...
  provider: signalfx
  signalFx:
    namespace:
      $ref: my-monioring-namespace.yml
    exporterUrl: http://signalfx-prometheus-exporter:9091/metrics
    targetFilterLabel: probe
```

then define an endpoint and reference the provider

```yaml
  ---
  $schema: /app-sre/app-1.yml

  ...

  endpoints:
  - name: my-probe-name
    description: Status
    url: https://my-service.com
    monitoring:
    - provider:
        $ref: my-signalfx-provider.yml

```

this will result in the setup of a `Probe` that will scrape for metrics with `probe=quay-io-status` labels,
where `probe` comes from `signalFx.targetFilterLabel` and `probe-name` comes from `endpoints.name`.

Signed-off-by: Gerd Oberlechner <goberlec@redhat.com>