explicite resource reference metadata in qontract-schema data files
Introduce a common type declaration `/common-1.json#/definitions/resourceref` that is compatible with `string` but represents a path to a resource.

```yaml
---
"$schema": /metaschema-1.json
version: '1.0'
type: object
...
properties:
  ...
  path:
    "$ref": "/common-1.json#/definitions/resourceref"
```

This information can be used in `qontract-validator` to verify resource references and enables fast fail for PR checks during bundle creation. At the same time it is compatible with current data.

Additionally add a property `isResourceRef = true` to the respective field in the graphql schema. This can leveraged during schema introspection.

During bundle creation, `resources.backrefs` will be filled with a list of data files that reference a resource.

References:
* design doc https://gitlab.cee.redhat.com/service/app-interface/-/blob/master/docs/app-sre/design-docs/qontract-schema-resource-references.md
* https://issues.redhat.com/browse/APPSRE-5629

Signed-off-by: Gerd Oberlechner <goberlec@redhat.com>