add resolveResource flag to graphql schema types (#186)
the resolveResource flag enables the resolution of resource ref fields during query time.
additionally, isResource is metadata enough to determine if a field is a resource. the nuance to determine if a field is a direct resource or just a reference to it, can be done with the field type, e.g. string vs Resource_v1

see https://github.com/app-sre/qontract-server/pull/147 for the qontract-server change

ref: https://issues.redhat.com/browse/APPSRE-5950

Signed-off-by: Gerd Oberlechner <goberlec@redhat.com>