add support to enable query support (#225)
related to https://issues.redhat.com/browse/APPSRE-6098

to speed up openshift-resources (and prometheus-rules-tester) early-exit, we need to skip as many resources as possible.

resources with dynamic content (generated based on graphql query results) must be checked, but only them.
with this field, we will know exactly the ones to check. we will add a validation that if the query() function is used, this option must be set to true.