resolvable resource references for openshift-resources
in order to improve performance for openshift-resources related integrations and their early-exit mode, their resources will be fetched inline along with the rest of the data instead of sequential reads.

it is crucial to align the rollout of this schema change with the qontract-reconcile integration using those fields (openshift-resources, openshift-routes, prometheus-rules-tester).

relates to https://issues.redhat.com/browse/APPSRE-6098

Signed-off-by: Gerd Oberlechner <goberlec@redhat.com>