cna assume-role resource (#305)
temporarily add support for the `example-aws-assumerole` CNA type.

```
  - provider: aws-assume-role
    identifier: test
    aws_assume_role:
      slug: test
      account:
        $ref: /aws/some-acc/account.yml
```

the role ARNs this and other kinds require, are stored in the `/aws/account-1.yml` schema. a default role ARN can be defined but module specific ones are supported as well. while we iterate we might drop support for the defaultRoleARN or revamp the way of providing module/context specific default parameters

```
---
$schema: /aws/account-1.yml

cna:
  defaultRoleARN: arn:aws:iam::xxx:role/RH-CNA
  moduleRoleARNS:
  - module: aws-rds
     arn: arn:aws:iam::xxx:role/RH-CNA-RDS
```

Signed-off-by: Gerd Oberlechner <goberlec@redhat.com>