[cna] RDS overrides as json with all optional
revert experiment where overrides and defaults are exactly the same schema. the result was that all fields in overrides and defaults needed to be optional so they can be used in both places without making overrides mandatory.

in this PR, overrides are now all optional, follow a schema in jsonschema but none in GQL where they are just JSON

Signed-off-by: Gerd Oberlechner <goberlec@redhat.com>