[change-types] implicit ownership for change-types (#360)
For certain changes, the natural approvers can be found in the datafiles themselves and would not require explicit ownership via roles. To reduce the need for such explicit ownership, we can make change-types implicitly applicable on changes.

The following example enables users to approve updates on their `public_gpg_key`s. The `jsonPathSelector` is the mechanism to find a reference to an approver, either a `/access/user-1.yml` or `/access/bot-1.yml` object. In this example the selector is `$.path` which finds the approver in the changed file itself.

```yaml
$schema: /app-interface/change-type-1.yml
...
contextSchema: /access/user-1.yml
changes:
- public_gpg_key

implicitOwnership:
- provider: jsonPath
  jsonPathSelector: $.path
```

https://issues.redhat.com/browse/APPSRE-6629

Signed-off-by: Gerd Oberlechner <goberlec@redhat.com>