[ocm] add mandatory orgId field
SA tokens usually have no orgId associated with them but we still rely on `ocm whoami` to find the org id to use when filtering for clusters. when we move forward with SAs that don't have an org association, we need another way to identify the org of an `/openshift/openshift-cluster-manager-1.yml`.

this schema changes makes the org association of an `/openshift/openshift-cluster-manager-1.yml` explicit by adding an `orgId` property to it.

attention: this is a mandatory new property, so a breaking schema change

https://issues.redhat.com/browse/APPSRE-7252

Signed-off-by: Gerd Oberlechner <goberlec@redhat.com>