[ocm] introduce schema to for an OCM environment
the current `/openshift/openshift-cluster-manager-1.yml` schema represents an OCM organization and not a instance of OCM. since we are adding more and more OCM organizations, this will lead to repeating metadata like URLs and credentials.

certain integrations will also start revolving around OCM instances and not organizations so having all of them as referenceable resources is a good next schema evolution.

https://issues.redhat.com/browse/APPSRE-7281

Signed-off-by: Gerd Oberlechner <goberlec@redhat.com>