[glitchtip-project-1.yml] add glitchtipId attribute (#491)
### What
Add an optional `glitchtip-project-1.projectId` attribute. 

### Why
In Glitchtip, a project is referenced by a `slug`; for convenience, the integration deviates this `slug` based on the `glitchtip-project-1.name` attribute. If a tenant wants to rename a project, the integration can't figure out the original `slug` anymore and will delete and recreate the project. To omit that, a tenant can manually specify the `slug` via `glitchtip-project-1.projectId ` to rename a project.