__identifier field on crossrefs to enable validation of context uniqueness in lists of crossref objects
the `__identifier` field is a field managed by qontract-validator when an object defines one or many `isContextUnique` fields. it holds a hash which is then used to figure out if an object is really unique within the context.

to enable the context uniqueness check also on lists of crossrefs, the crossref jsonschema type needs to declare it so that qontract-bundler can add it and qontract-validator can still validate the schema.

relates to https://github.com/app-sre/qontract-validator/pull/58

Signed-off-by: Gerd Oberlechner <goberlec@redhat.com>