APPSRE-9922: Add network claim schema
As it stands, this will not validate because the
`  - { name: inUseBy, type: string, isRequired: false }`
is wrong; it should pair with the definition to be properly supportive
of the polymorphic nature of allowing multiple things to claim a
network. On the topic of claiming a network, vpc-1 and app-1 are used as
placeholders and do not necessarily represent the final configuration.

Signed-off-by: Lisa Seelye <lisa@users.noreply.github.com>