Globally prefer podman and better cross platform support
There were some targets in the Makefile that were hardcoded to use the
docker binary. They have been changed to use `$(CONTAINER_ENGINE)`,
which will prefer podman and fall back to docker.

In order to better support non-Linux platforms, some auto-detection is
done for the OS and if Linux is detected, container volumes will be
mounted with `:z` to facilicate SELinux labels (which are not supported
on MacOS with Podman). There are two ways to override the
auto-detection. To pretend to _not_ be Linux: `make UNAME=Unknown` (or
whichever platform). To remain Linux (for example) but disable SELinux
mount: `make SELINUX_MOUNT_CHAR=`

Signed-off-by: Lisa Seelye <lisa@users.noreply.github.com>