prometheusrule: Modify required alert annotation schema
In upstream [monitoring
mixins](https://monitoring.mixins.dev/prometheus/) as well as [OpenShift
alerts best
practices](https://github.com/openshift/enhancements/blob/master/enhancements/monitoring/alerting-consistency.md#style-guide)
the annotations that must be included are `description` and `summary`.
`message` is usually not an annotation that is present.

This commit modifies the schema to be inline with these.

Signed-off-by: Saswata Mukherjee <saswataminsta@yahoo.com>